docs: document transport module and harden transport system
