+++ packages/engine/src/simulation/citySimulationEngine.ts
@@ -3,7 +3,7 @@ import { RoadNetworkSystem } from './roadNetwork';
 import { TrafficSimulationSystem } from './trafficSimulation';
 import { ZoningSystem, ZoneType } from './zoningSystem';
 import { CityServicesSystem, ServiceType } from './cityServices';
-import { PublicTransportSystem } from './publicTransport';
+import { PublicTransportSystem } from './transport/transportSystem';
 
 export interface CityConfig {
   gridWidth: number;
